requirements.txt
pandas
ollama
openai
python-dotenv
google-generativeai
openai-agents
faiss-cpu
crewai
crewai-tools
mailersend
autogen-agentchat
autogen-ext[openai]
mcp
langgraph 
langchain-community 
langchain-huggingface 
langchain-core 
langchain-text-splitters 
langgraph-checkpoint-sqlite
sentence-transformers
pypdf
pydantic
chromadb
langchain-tavily
tavily-python
autogen==0.9.9
autogen-core==0.7.4
asyncio